requirements.txt

pandas==2.1.0
streamlit==1.36.0
python-docx